All codes/Handoffs Agent/pyproject.toml
[project]
name = "handoffs-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.17.0",
    "python-dotenv>=1.2.2",
]
